requirements.txt
Flask==0.12.2
flask-restful==0.3.6
markdown==2.6.11
pytest==3.8.0
requests==2.23.0
bcrypt==3.1.7
pyjwt==1.7.1